refactor(supervisor): rename supervisor model
