+++ flyve-mdm-ios/AppDelegate.swift
@@ -125,7 +125,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             supervisorObject["support_website"] = invitation[5]
             supervisorObject["support_email"] = invitation[6]
             
-            let supervisor = SupervisorViewModel(supervisor: Supervisor(data: supervisorObject as [String: AnyObject]))
+            let supervisor = SupervisorViewModel(supervisor: SupervisorModel(data: supervisorObject as [String: AnyObject]))
             
             setStorage(value: invitationObject as AnyObject, key: "deeplink")
             setStorage(value: supervisor as AnyObject, key: "supervisor")
