+++ flyve-mdm-ios/Enroll.swift
@@ -1,7 +1,7 @@
 /*
  *   Copyright © 2017 Teclib. All rights reserved.
  *
- * SupervisorViewModel.swift is part of flyve-mdm-ios
+ * Enroll.swift is part of flyve-mdm-ios
  *
  * flyve-mdm-ios is a subproject of Flyve MDM. Flyve MDM is a mobile
  * device management software.
@@ -26,31 +26,3 @@
  */
 
 import Foundation
-
-class SupervisorViewModel {
-    private let supervisor: SupervisorModel
-    
-    var name: String {
-        return supervisor.name
-    }
-    
-    var email: String {
-        return supervisor.email
-    }
-    
-    var phone: String {
-        return supervisor.phone
-    }
-    
-    var website: String {
-        return supervisor.website
-    }
-    
-    var picture: String {
-        return supervisor.picture
-    }
-
-    init(supervisor: SupervisorModel) {
-        self.supervisor = supervisor
-    }
-}
